refactor: merge tenant usage library to quota library
