+++ packages/core/src/test-utils/quota.ts
@@ -1,11 +1,35 @@
-import { type QuotaLibrary } from '#src/libraries/quota.js';
+import type { CloudConnectionLibrary } from '#src/libraries/cloud-connection.js';
+import type { ConnectorLibrary } from '#src/libraries/connector.js';
+import { QuotaLibrary } from '#src/libraries/quota.js';
+import type { SubscriptionLibrary } from '#src/libraries/subscription.js';
+
+import { createMockCommonQueryMethods } from './query.js';
 
 const { jest } = import.meta;
 
+class MockQuotaLibrary extends QuotaLibrary {
+  constructor() {
+    super(
+      createMockCommonQueryMethods(),
+      // eslint-disable-next-line no-restricted-syntax
+      {} as ConnectorLibrary,
+      'mock-tenant-id',
+      // eslint-disable-next-line no-restricted-syntax
+      {} as CloudConnectionLibrary,
+      // eslint-disable-next-line no-restricted-syntax
+      {} as SubscriptionLibrary
+    );
+  }
+
+  // Override the methods with jest mock functions
+  // eslint-disable-next-line @typescript-eslint/member-ordering
+  guardTenantUsageByKey = jest.fn();
+  // eslint-disable-next-line @typescript-eslint/member-ordering
+  guardEntityScopesUsage = jest.fn();
+  // eslint-disable-next-line @typescript-eslint/member-ordering
+  reportSubscriptionUpdatesUsage = jest.fn();
+}
+
 export const createMockQuotaLibrary = (): QuotaLibrary => {
-  return {
-    guardTenantUsageByKey: jest.fn(),
-    guardEntityScopesUsage: jest.fn(),
-    reportSubscriptionUpdatesUsage: jest.fn(),
-  };
+  return new MockQuotaLibrary();
 };
